WordleFrame Arduino firmware
Code to drive a Neopixel / E-ink based display of Wordle scores.  Data
comes from an adafruit.io MQTT topic and is shown on the hardware.  The
E-ink shows user name and score, while the NeoPixels show the guesses.

Hardware is based on the Adafruit MagTag arduino and a custom built
PCB for the LEDs.